requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.2.0
numpy==1.26.4